fix(ams-ui): renamed table component file
